+++ CHANGES.rst
@@ -25,6 +25,9 @@ Documentation
 *************
 
 - #230: Document return value of the client's ensure_path method.
+- #222: Document that committed on the transaction is used to ensure only one
+  commit and is not an indicator of whether operations in the transaction
+  returned desired results.
 
 2.0 (2014-06-19)
 ----------------